Adding a wire protocol
HTTP API is great for simplicity and testing, but you lose add a *lot* of overhead in requests. On each request, you're adding a bunch of HTTP headers that aren't necessary to the DB's functionality.

A wire protocol is therefore necessary if you want your database benchmarks to be competitive at all. It highly decreases the number of bytes per request and the amount of latency for each request, since you can have a long lived connection.